Fixed a bug exhibited in MapReduce::copy() (but it would have affected other functions as well).
The MapReduce instance counter grows and shrinks as MapReduce objects are
instantiated and destroyed.  The shrinking is the problem:  it allowed
several MapReduce objects to have the same instance number and, thus,
use the same filenames for out-of-core files.   This really confused the
MapReduce::copy() function when, through bad luck, both the original MR 
object and the copy used the same filenames!

We need the counter that grows and shrinks to determine when to call
MPI_Finalize.  But we also need a counter that doesn't shrink to keep the 
filenames consistent.  I added a non-shrinking counter.  As long as 
we don't instantiate more than 2^31 MR objects, this counter will be OK.




git-svn-id: svn+ssh://software-srn.sandia.gov/svn/mapreduce/trunk@287 3fc257bb-8550-0410-81b3-a5c103da83f7